Refactor page content to be loaded by vite
