+++ src/entry-server.js
@@ -1,7 +1,11 @@
+import van from 'mini-van-plate/van-plate'
 import javascriptLogo from './javascript.svg'
+import page from './page'
 
 export function render() {
+  const content = page({ van })
   const html = `
+    ${van.html(content)}
     <div>
       <a href="https://vitejs.dev" target="_blank">
         <img src="/vite.svg" class="logo" alt="Vite logo" />
